Added Using Fragment and Computed Fields
Added Using Fragment and Computed Fields

1. Using Fragment
- When comment need to be confirmed, we should be able to access User info of comments. In order to confirm user info (username, etc...), we use $fragment() method and file of fragments.js defined info of fragment.

2. Computed Fields
We need to fields and should be able to access them. they aren't not in DB but just created in resolver file by calculating something. In resolver file, they can be written next Query or Mutation. And "parent" argument can be used to calculate result value for computed field. And then return them.